move tensor input api
