+++ src/aioway/modes/hists.py
@@ -10,9 +10,9 @@ from collections import abc as cabc
 
 import torch
 
-from aioway._iters import TensorInput
 from aioway._thunks import Thunk
 from aioway._utils import (
+    TensorInput,
     find_nested_tensors,
     is_leaf_has_grad,
     replace_tensors_with_attr,
